Fix assignment of grade to non-existent student
This issue is most likely to show up when grading
by group and a student in the group has dropped the
class. As the group grade assignment loops over
students and runs `Grader.assignGradeStudent`,
if one student is no longer present in the
gradebook (dropped or switched sections), an
error would occur where the dataframe lookup
returns an empty dataframe, which becomes an
IndexError when accessing the zeroth element
with `.values[0]`. Thus, add a check for an
empty dataframe before accessing the current
grade to check it for ABS or EX grades.